fix(api): address review comments on additionalAdvertisePeerUrls
- Remove redundant regex match from Etcd-level CEL index bounds rule
  (pattern already enforced at field level via +kubebuilder:validation:Pattern)
- Drop redundant && !u.startsWith(...) from TLS scheme CEL rules
  (startsWith('https://') and startsWith('http://') are mutually exclusive)
- Remove +kubebuilder:validation:MinLength=1 from MemberName
  (regex pattern already enforces minimum length)
- Clean up godoc on AdditionalAdvertisePeerURLs and MemberName fields;
  remove inaccurate "silently ignored" wording
- Replace builder.WriteString(fmt.Sprintf(...)) with fmt.Fprintf
  per staticcheck QF1012
- Move additionalAdvertisePeerUrls CEL validation tests into
  specetcd_test.go alongside other spec.etcd field tests
